Extract the Gleam Time component
- Move wall-clock providers into geam-time.
- Preserve caller-owned source and stdlib composition.
- Keep the existing root Time facade unchanged.
